2008-07-02  Pavel Roskin  <proski@gnu.org>
	* util/grub-editenv.c (cmd_info): Cast argument to long
	explicitly.  ptrdiff_t reduces to int on i386.

+++ util/grub-editenv.c
@@ -141,7 +141,7 @@ open_envblk_file (char *name)
 static void
 cmd_info (void)
 {
-  printf ("Envblk offset: %ld\n", envblk->data - buffer);
+  printf ("Envblk offset: %ld\n", (long) (envblk->data - buffer));
   printf ("Envblk length: %d\n", envblk->length);
 }
